SQL util extracted
