+++ lib/windshaft/renderers/pg-mvt/factory.js
@@ -4,7 +4,7 @@ var layersFilter = require('../../utils/layer_filter');
 
 var RendererParams = require('../renderer_params');
 
-var sql = require('./sql');
+var sql = require('../../utils/sql');
 var Renderer = require('./renderer');
 var BaseAdaptor = require('../base_adaptor');
