+++ lib/windshaft/renderers/torque/factory.js
@@ -8,7 +8,7 @@ var format = require('../../utils/format');
 
 var RendererParams = require('../renderer_params');
 
-var sql = require('./sql');
+var sql = require('../../utils/sql');
 var Renderer = require('./renderer');
 var PSQLAdaptor = require('./psql_adaptor');
 var PngRenderer = require('./png_renderer');
