+++ lib/windshaft/utils/sql.js
@@ -1,4 +1,4 @@
-var debug = require('debug')('windshaft:torque');
+var debug = require('debug')('windshaft:utils');
 
 function sql(dbParams, SQLClass) {
     // TODO: cache class instances by dbParams/sqlClass
